front refactored into componnets v 1.0
